requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
Authlib==1.3.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.110
botocore==1.34.110
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.5.0
cryptography==42.0.7
decorator==4.4.2
distro==1.9.0
distyll-info==0.2.4
exceptiongroup==1.2.1
fastavro==1.9.4
filelock==3.14.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.0
grpcio-health-checking==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.1
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moviepy==1.0.3
mutagen==1.47.0
numpy==1.26.4
ollama==0.2.0
openai==1.30.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proglog==0.1.10
protobuf==4.25.3
pyarrow==16.1.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pypdf==4.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
streamlit==1.34.0
tenacity==8.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
types-requests==2.32.0.20240521
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.1
weaviate-client==4.6.3
websockets==12.0
yt-dlp==2023.12.30
